Add files via upload
Hi everyone, In the next few lines we will be creating a working database using the language SQL. Not only that but we will be going to take look at some of the examples queries to fetch data from our database. We will be executing this code in MySQL, If you are using any other Database Management Software make sure to check with the syntax as some of them might have slight changes.